add scripts
